feat(admin): move sightings table to /admin/sichtungen with redirect
