+++ src/lib/config/adminNav.test.ts
@@ -24,18 +24,18 @@ describe('istAdminPfad', () => {
 });
 
 describe('aktiverAdminBereich', () => {
-	it('ordnet die drei Bereiche ihrem eigenen Pfad zu', () => {
+	it('ordnet die vier Bereiche ihrem eigenen Pfad zu', () => {
 		for (const bereich of ADMIN_BEREICHE) {
 			expect(aktiverAdminBereich(bereich.href)).toBe(bereich.href);
 		}
 	});
 
-	it('rechnet Detail- und Referenzseiten der Sichtungsverwaltung zu', () => {
+	it('rechnet Detail- und Referenzseiten der Sichtungstabelle zu', () => {
 		/* Ohne diese Zuordnung verlöre die Unternavigation beim Öffnen einer
 		   Sichtung ohne erkennbaren Grund ihre Markierung. */
-		expect(aktiverAdminBereich('/admin/1234')).toBe('/admin');
-		expect(aktiverAdminBereich('/admin/1234/edit')).toBe('/admin');
-		expect(aktiverAdminBereich('/admin/ref/AB-12')).toBe('/admin');
+		expect(aktiverAdminBereich('/admin/1234')).toBe('/admin/sichtungen');
+		expect(aktiverAdminBereich('/admin/1234/edit')).toBe('/admin/sichtungen');
+		expect(aktiverAdminBereich('/admin/ref/AB-12')).toBe('/admin/sichtungen');
 	});
 
 	it('markiert auf Unterseiten der Statistiken und Einstellungen den richtigen Bereich', () => {
@@ -51,5 +51,34 @@ describe('aktiverAdminBereich', () => {
 		expect(aktiverAdminBereich('/')).toBeNull();
 		expect(aktiverAdminBereich('/map')).toBeNull();
 		expect(aktiverAdminBereich('/administration')).toBeNull();
+	});
+});
+
+describe('Admin-Navigation mit Eingangsseite', () => {
+	it('führt Eingang und Sichtungen als getrennte Bereiche', () => {
+		expect(ADMIN_BEREICHE.map((b) => b.href)).toEqual([
+			'/admin',
+			'/admin/sichtungen',
+			'/admin/statistics',
+			'/admin/settings'
+		]);
+	});
+
+	it('/admin markiert den Eingang', () => {
+		expect(aktiverAdminBereich('/admin')).toBe('/admin');
+	});
+
+	it('/admin/sichtungen markiert die Tabelle', () => {
+		expect(aktiverAdminBereich('/admin/sichtungen')).toBe('/admin/sichtungen');
+	});
+
+	it('Detail-, Bearbeiten- und Referenzseiten gehören zur Tabelle', () => {
+		expect(aktiverAdminBereich('/admin/123')).toBe('/admin/sichtungen');
+		expect(aktiverAdminBereich('/admin/123/edit')).toBe('/admin/sichtungen');
+		expect(aktiverAdminBereich('/admin/ref/abc')).toBe('/admin/sichtungen');
+	});
+
+	it('/admin/docs bleibt ohne aktiven Bereich', () => {
+		expect(aktiverAdminBereich('/admin/docs')).toBeNull();
 	});
 });
