+++ src/routes/admin/sichtungen/dateParam.ts
